added functionality with validation      * Admin konczy mecz (zwyciestwo 2)      * ...      * Uzytkownik A - nic nie dostaje      * Uzytkownik B - dostaje kase, 1 pkt      * ...

+++ src/main/java/com/dominik/typer/model/Bet.java
@@ -6,6 +6,8 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 @AllArgsConstructor
@@ -16,5 +18,5 @@ public class Bet {
     Integer userId;
     Integer pointsFromBetEvent;
     BetType betType;
-    Integer betAmount;
+    BigDecimal betAmount;
 }
